- wxbRestorePanel & wxbHistoryTextCtrl : Fixed unsigned comparisons.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1477 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/wx-console/wxbhistorytextctrl.cpp
@@ -50,7 +50,7 @@ void wxbHistoryTextCtrl::HistoryAdd(wxString cmd) {
 void wxbHistoryTextCtrl::OnKeyUp(wxKeyEvent& event) {
    if (event.m_keyCode == WXK_UP) {
       if (index > 0) {
-         if (index == (history.Count()-1)) {
+         if (index == ((int)history.Count()-1)) {
             history[index] = GetValue();
          }
          index--;
@@ -59,7 +59,7 @@ void wxbHistoryTextCtrl::OnKeyUp(wxKeyEvent& event) {
       }
    }
    else if (event.m_keyCode == WXK_DOWN) {
-      if (index < (history.Count()-1)) {
+      if (index < ((int)history.Count()-1)) {
          index++;
          SetValue(history[index]);
          SetInsertionPointEnd();
